python/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==40.0.2
Jinja2==3.0.3
junos-eznc==2.6.8
lxml==5.3.0
MarkupSafe==2.0.1
ncclient==0.6.13
paramiko==3.5.0
pkg_resources==0.0.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.1.4
pyserial==3.5
PyYAML==6.0.1
scp==0.15.0
six==1.17.0
transitions==0.9.2
yamlordereddictloader==0.4.2
